Add socket.io to `server.js`
Configured `server.js` to use socket.io.  Using the
`app.set()` method introduced in Express 4 so that
I can get access to the `io` object inside of the
routes file (otherwise it would be a circular
require mess).  To access `io` in a route, call
`app.get(io)`.  Also created a listener for socket
connections which will transmit the socket id back
to the client.  This allows the client to send the
socket id with requests via query params making it
possible to use it to access the socket and then
send broadcast emits instead of emitting messages to
all sockets.